added delete button --EventsListRow.js
