+++ src/components/Admin/EventsListRow.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import Button from 'react-bootstrap/Button';
 
 const EventsListRow = (props) => {
     return (
@@ -11,6 +12,9 @@ const EventsListRow = (props) => {
             </td>
             <td>
                 {props.description}
+            </td>
+            <td style={{'text-align': 'center'}}>
+                <Button variant="danger" className='m-auto' style={{'padding': '1px 10px', 'margin': 'auto'}} onClick={() => props.handleEventRemove(props.id, props.title)} >Delete</Button>
             </td>
         </tr>
     )
